requirements.txt
Django==2.2.16
django-filter
Pillow
python-dateutil
Unipath
django-simple-history
django-widget-tweaks
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-extensions
django-import-export==2.0.1
django-storages==1.9.1
boto3==1.12.39
requests==2.22.0
mysqlclient==1.4.6
requests==2.22.0
django-cors-headers
django-admin-rangefilter
django-3-jet
pytest
